more fields mapped in xml sidecar and methods deprecated in mh_properties
